% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.utils.R
\name{ablack}
\alias{ablack}
\title{Quick Helper function to generate an opaque Black Color Value.}
\usage{
ablack(alpha = 0.2)
}
\arguments{
\item{alpha}{Alpha value (between 0 and 1) applied to black.}
}
\description{
Quick Helper function to generate an opaque Black Color Value.
}
