% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.utils.R
\name{acol}
\alias{acol}
\title{Apply Alpha Channel to Colors}
\usage{
acol(x, alpha = 0.2)
}
\arguments{
\item{x}{A color or a vector of colors. Can be color names or hex codes.}

\item{alpha}{A numeric value (between 0 and 1) or a vector of values representing the
transparency level(s) to apply. Must be the same length as `x` if vectorized.}
}
\value{
A character vector of colors with alpha applied, in the same length as `x`.
}
\description{
Adds an alpha transparency to one or more colors.
}
