% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_utils.R
\name{approx.density.func}
\alias{approx.density.func}
\title{Function to roughly estimate a kernel density estimate from a sample}
\usage{
approx.density.func(
  dsamp,
  bounds = NULL,
  num.points = 1000,
  bw = "nrd0",
  method = "monoH.FC",
  elim.zero.dsQ = T,
  tol = 1e-19,
  eps = .Machine$longdouble.eps,
  plotQ = F,
  bre = NULL,
  printQ = F
)
}
\arguments{
\item{bounds}{Left and right bounds on distribution. If NULL, min/max of sample used}

\item{num.points}{Length of support domain (x-axis) to be used for distribution}

\item{bw}{Bandwidth to be used for density fits. Cf. density()}

\item{method}{Fit method for density. Cf. density()}

\item{tol}{Minimum fit density value. Below tol, density values will be replaced with eps. If NULL, zero density values are kept.}

\item{eps}{Replacement density value for densities below tol}

\item{plotQ}{Plot flag}

\item{bre}{Parameter bre for hist(). Defaults to length(pretty(dsamp))}

\item{printQ}{Print flag}

\item{dat.list}{Samples from two distributions. Send in as a list in case samples are not the same length}
}
\value{
The function will XX
}
\description{
Function to roughly estimate a kernel density estimate from a sample. Useful with function in entropy package.
}
