% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_utils.R
\name{approx.KL.divergence}
\alias{approx.KL.divergence}
\title{The function to bin and get frequencies for data}
\usage{
approx.KL.divergence(
  dat.list,
  bounds = NULL,
  num.points = 1000,
  bw = "nrd0",
  log.base = exp(1),
  renormalizeQ = T,
  tol = 9e-19,
  eps = .Machine$longdouble.eps,
  subdivisions = 2000,
  plotQ = F,
  printQ = F
)
}
\arguments{
\item{dat.list}{Samples from two distributions. Send in as a list in case samples are not the same length}

\item{bounds}{Left and right bounds on distributions. If NULL, min/max of samples used}

\item{num.points}{Length of common support domain (x-axis) to be used for both distributions}

\item{bw}{Bandwidth to be used for density fits. Cf. density()}

\item{log.base}{Log base to be used in KL formula}

\item{renormalizeQ}{Whether or not to re-normalize density fits so that they are normalized over the specified bounds}

\item{tol}{Minimum fit density value. Below tol, density values will be replaced with eps}

\item{eps}{Replacement density value for densities below tol}

\item{subdivisions}{Maximum number of subintervals for integrate()}

\item{plotQ}{Plot flag}

\item{printQ}{Print flag}
}
\value{
The function will XX
}
\description{
Handy for use with discrete entropy and related functions
}
