% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{autocorrelation.plots}
\alias{autocorrelation.plots}
\title{XXXX}
\usage{
autocorrelation.plots(chains.array, pars = NULL)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
XXXX
}
