% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.utils.R
\name{bu.color}
\alias{bu.color}
\title{Retun bayesutil's Default Colors}
\usage{
bu.color(i = 1, alpha = 1)
}
\arguments{
\item{i}{Color integer index, between 1 - 9.}

\item{alpha}{Alpha color value to apply if needed. Defaults to 1.}
}
\description{
Retun bayesutil's Default Colors
}
