% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.utils.R
\name{ccol}
\alias{ccol}
\title{Produce a Simple RGB Complement Color given an Input Color}
\usage{
ccol(x, rot = 1)
}
\arguments{
\item{x}{A color, can be a color name or hex code.}

\item{rot}{A numeric value. Can be either 1 or 2. Then the channels are rotated by that number.}
}
\value{
A character vector of complement color `x`.
}
\description{
Generate an RGB complement by transposing channel values.
}
