% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{chain.idx}
\alias{chain.idx}
\title{Generate Chain Indices}
\usage{
chain.idx(fit_obj)
}
\arguments{
\item{fit_obj}{A CmdStanFit object or similar.}
}
\value{
An integer vector indicating the chain each draw belongs to.
}
\description{
Returns the chain index for each draw in a posterior sample.
}
