% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help.R
\docType{data}
\name{clean.me.up.chain}
\alias{clean.me.up.chain}
\title{Unclean MCMC chain}
\format{
An object of class \code{data.frame} with 9987 rows and 1 columns.
}
\usage{
data(clean.me.up.chain)
}
\description{
An uncleaned up MCMC chain. Useful for exploring computation of convergence diagnostics and cleanup
methodologies.
}
\examples{
data(clean.me.up.chain)
}
\keyword{datasets}
