% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal.R
\docType{data}
\name{coal}
\alias{coal}
\title{Coal Data}
\format{
A data frame with 112 rows and 2 columns:
\describe{
\item{year}{Year.}
\item{disasters}{Number of mining disasters.}
}
}
\usage{
data(coal)
}
\description{
Counts for the number of coal mine disasters over a
112-year period (1851 to 1962) in the United Kingdom.
}
