% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_20210307.R
\docType{data}
\name{covid_20210307}
\alias{covid_20210307}
\title{COVID-19 statistics}
\format{
A data frame with 50 rows and 7 variables \describe{
\item{state_name}{state name} \item{state_abb}{state name abbreviation}
\item{deaths}{number of confirmed and probable COVID-19 deaths}
\item{cases}{number of confirmed and probable COVID-19 cases}
\item{population}{population 5-year stimate from the 2017 American
Community Survey}
\item{income}{median income (U.S. dollars) 5-year
estimate from the 2017 American Community Survey}
\item{hs}{estimated
percentage of population > 25 years old with a high school diploma based on
the 2010 Census}
\item{bs}{estimated percentage of population > 25 years
old with a Bachelor's degree based on the 2010 Census}
\item{vote_diff_2020}{difference between percentage of voters who voted for
Joseph Biden versus Donald Trump in the 2020 election. This is (total Biden
votes - total Trump votes) * 100 / (total Biden votes + total Trump votes)}} }
}
\source{
\url{https://covidtracking.com/data/download} (Accessed 2020-03-08) and \url{https://www.census.gov}
MIT Election Data and Science Lab, 2017, "U.S. President 1976–2020", \url{https://doi.org/10.7910/DVN/42MVDX}, Harvard Dataverse, V6
}
\usage{
data(covid_20210307)
}
\description{
This data set contains the number of confirmed and probable COVID-19 deaths and cases
in U.S. states through March 7th, 2021 along with some very basic
demographic information. While the data are taken from official sources,
they are only intended for demonstration purposes.
}
