% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_dec4.R
\docType{data}
\name{covid_dec4}
\alias{covid_dec4}
\title{COVID-19 deaths}
\format{
A data frame with 50 rows and 7 variables
\describe{
\item{state_name}{state name}
\item{state_abb}{state name abbreviation}
\item{deaths}{confirmed number of COVID-19 deaths}
\item{population}{population e5-year stimate from the 2017 American Community Survey}
\item{income}{median income (U.S. dollars) 5-year estimate from the 2017 American Community Survey}
\item{hs}{estimated percentage of population > 25 years old with a high school diploma based on the 2010 Census}
\item{bs}{estimated percentage of population > 25 years old with a Bacherlor's degree based on the 2010 Census}
}
}
\source{
\url{https://covidtracking.com/data/download} (Accessed 2020-12-04) and \url{https://www.census.gov}
}
\usage{
data(covid_dec4)
}
\description{
This data set contains the number of COVID-19 deaths
in U.S. states through December 4th, 2020 along with some very basic
demographic information. While the data are taken from official sources,
they are only intended for demonstration purposes.
}
