% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{dens.at}
\alias{dens.at}
\title{Evaluate Density at Specific Locations}
\usage{
dens.at(d, x.at)
}
\arguments{
\item{d}{A density object created by [dens()].}

\item{x.at}{Numeric vector of x-locations at which to evaluate the density.}
}
\value{
A numeric vector of density values.
}
\description{
Returns the approximate density at specified x values.
}
