% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.R
\name{dom}
\alias{dom}
\title{dom}
\usage{
dom(pot.table)
}
\arguments{
\item{pot.table}{A potential (or node probability) table. Should be of class parray or array.}
}
\value{
The domain of random variates over which the bale is constructed.
}
\description{
Spit out domain of a CPT or POT.
}
\details{
Spit out domain of a CPT or POT:
}
\examples{
XXXX
}
