% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{draws.varnames}
\alias{draws.varnames}
\title{Extract Variable Names from Draws}
\usage{
draws.varnames(draws, ..., unique.names = TRUE)
}
\arguments{
\item{draws}{A list of posterior samples.}

\item{...}{Additional draw lists to include.}

\item{unique.names}{Logical; if `TRUE`, returns unique variable names.}
}
\value{
A character vector of variable names.
}
\description{
Retrieves a flat list of variable names including array indices.
}
