% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_utils.R
\name{extract.log.lik}
\alias{extract.log.lik}
\title{Function to extract log likelihoods form Stan or JAGS fit object}
\usage{
\method{extract}{log.lik}(fit.obj, parameter_name = "log_lik", merge_chains = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Function to extract log likelihoods form Stan or JAGS fit object. Mirror functionality
of extract_log_lik in the loo package. Returns object suitable for use with loo package.
}
\details{
The function is XXXXXXX
}
