% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_utils.R
\name{extract.params}
\alias{extract.params}
\title{XXXXX}
\usage{
extract.params(
  fit.obj,
  params.vec = NULL,
  by.chainQ = F,
  as.data.frameQ = F,
  as.matrixQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXXX
}
\details{
The function is XXXXXXX
}
