% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract Posterior Samples}
\usage{
extract_samples(fit_obj, chain.subset = 0)
}
\arguments{
\item{fit_obj}{A fitted model object. Can be of class `stanreg`, `CmdStanFit`, or similar.}

\item{chain.subset}{An integer indicating the chain to extract. Use 0 for all chains.}
}
\value{
A list of posterior samples, one element per variable, with class `"mcmc.draws"`.
}
\description{
Extracts posterior samples from a fitted model object (Stan or similar).
}
