% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{extract_samples_from_rstan}
\alias{extract_samples_from_rstan}
\title{Extract Samples from rstan Object}
\usage{
extract_samples_from_rstan(fit_obj, chain.subset = 0)
}
\arguments{
\item{fit_obj}{An object of class `stanreg`.}

\item{chain.subset}{An integer indicating the chain to extract. Use 0 for all chains.}
}
\value{
A list of samples with class `"mcmc.draws"`.
}
\description{
Extracts posterior samples from a `stanreg` object from the rstan package.
}
