% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{get.bin.freqs}
\alias{get.bin.freqs}
\title{The function to bin and get frequencies for data}
\usage{
get.bin.freqs(dat.list, num.bins, countsQ = F)
}
\arguments{
\item{dat.list}{The data. Send data in as a list in case data vectors are not the same length}

\item{countsQ}{Return bin counts instead of frequencies}

\item{num.bin}{How many bins to use}
}
\value{
The function will XX
}
\description{
Handy for use with discrete entropy and related functions
}
