% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.file_utils.R
\name{get.mfp}
\alias{get.mfp}
\title{Get path to the specified JAGS or Stan model file contained in bayesutils}
\usage{
get.mfp(model.file.name)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Get path to the specified JAGS or Stan model file contained in bayesutils
}
