% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_file}
\alias{get_file}
\title{Get a model file or download and load a script file}
\usage{
get_file(file.name, url_head = NULL, download_location = NULL)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Get a model file or download and load a script file. Combines functionality of
get.script() and open_mf() into one function for easier use.
}
