% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplot.densities}
\alias{ggplot.densities}
\title{Produce (Credible / Percentile) Interval Plot of for a Set of Samples.}
\usage{
ggplot.densities(
  xs = NULL,
  dens_list = NULL,
  xnames = NULL,
  pi_lvls = c(0.5, 0.9, 0.99),
  pi_fun = PI,
  pi_lab_digits = 2,
  ymax = NULL,
  xlim = NULL,
  show_densline = T,
  show_pi_segments = T,
  show_pi_labels = T,
  show_mean = T,
  show_yaxis = T,
  colors = bu.color(7),
  col_mean = "white",
  col_pi_labels = "grey20",
  fontface_pi_labels = "italic",
  pi_label_formatter = round.format,
  ylab = "density",
  xlab = "x",
  ggtheme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{xs}{List of numeric sample vectors.}

\item{dens_list}{Optional list of density objects, i.e. obtained by the density function.}

\item{xnames}{Sample names. Could also be supplied by naming xs and/or dens_list.}

\item{pi_lvls}{Percentile levels for which interval bounds are plotted}

\item{pi_fun}{Function for computing intervals. Could be PI, HDI or a custom function.}

\item{pi_lab_digits}{Number of digits to show when interval labels are printed}

\item{ymax}{Upper limit of y axis.}

\item{xlim}{Lower and upper limit of x axis.}

\item{show_densline}{Should kernel density line be included.}

\item{show_pi_labels}{Should numeric labels for interval bounds be included.}

\item{show_mean}{Add the mean to the intervals visualization.}

\item{show_yaxis}{Hide y axis if necessary.}

\item{colors}{Colors to use to plot individual sample data.}

\item{col_mean}{Mean point color.}

\item{col_pi_labels}{Color of interval bound labels.}

\item{pi_label_formatter}{Function for formatting interval bound labels. Default generates scientific numerical notations when bound labels would be rounded to zero, as controlled by pi_lab_digits.}

\item{ylab}{Label of y axis.}

\item{xlab}{Label of x axis.}

\item{ggtheme}{ggplot2 theme to apply.}

\item{show_pi_segmenets}{Should (credible) percentile intervals be included to the plot.}

\item{font_face_pi_labels}{Label font face as defined by ggplot2. Defaults to italic.}
}
\description{
Given a list of numeric sample vectors, a plot of density estimates are produced.
Visualizations of (credible) percentile intervals are returned as well.
}
