% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplot.pi}
\alias{ggplot.pi}
\title{Produce (Credible / Percentile) Interval Plot}
\usage{
ggplot.pi(
  x,
  pi_lvls = c(0.5, 0.9, 0.99),
  pi_fun = PI,
  show_pi_labels = T,
  pi_lab_digits = 2,
  col_pi_segment_base = "grey20",
  col_pi_labels = "grey20",
  fontface_pi_labels = "italic",
  pi_label_formatter = round.format,
  xlab = "x",
  xlim = NULL
)
}
\arguments{
\item{x}{Numeric vector of sample values}

\item{pi_lvls}{Percentile levels for which interval bounds are plotted}

\item{pi_fun}{Function for computing intervals. Could be PI, HDI or a custom function.}

\item{show_pi_labels}{Should numeric labels for interval bounds be included}

\item{pi_lab_digits}{Number of digits to show when interval labels are printed}

\item{col_pi_segment_base}{Color of the interval segment. Opaqueness is added to this color with increasing percentile levels.}

\item{col_pi_labels}{Color of interval bound labels.}

\item{pi_label_formatter}{Function for formatting interval bound labels. Default generates scientific numerical notations when bound labels would be rounded to zero, as controlled by pi_lab_digits.}

\item{xlab}{Label of x axis.}

\item{xlim}{Lower and upper limit of x axis.}

\item{font_face_pi_labels}{Label font face as defined by ggplot2. Defaults to italic.}
}
\description{
Given a numeric vector x of some sample, a forest plot-like visualization
is produced, showing intervals of interest
}
