% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_utils.R
\name{HPDI}
\alias{HPDI}
\title{HPDI}
\usage{
HPDI(param.samp, credibility)
}
\arguments{
\item{pot.table}{A potential (or node probability) table. Should be of class parray or array.}
}
\value{
The domain of random variates over which the bale is constructed.
}
\description{
Highest posterior density interval
}
\details{
Wrapper for coda's HPDinterval function, so usuers don't have to convert sampling output
to coda mcmc format.
}
\examples{
XXXX
}
