% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igdata.R
\docType{data}
\name{igdata}
\alias{igdata}
\title{Inverse-Gamma Data}
\format{
A vector with 100 values.
}
\usage{
data(igdata)
}
\description{
This data set contains 100 observations
drawn from an Inverse-Gamma(shape = 2, scale = 3)
distribution.

The data can be used to implement an MCMC algorithm
to approximate the posterior distribution of the \code{shape}
and \code{scale} parameters of the distribution.
}
