% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvChisquare.R
\name{InvChiSquare}
\alias{InvChiSquare}
\alias{rinvchisq}
\alias{dinvchisq}
\alias{pinvchisq}
\alias{qinvchisq}
\title{The (scaled) Inverse Chi-squared Distribution}
\usage{
rinvchisq(n, df, scale = 1/df)

dinvchisq(x, df, scale = 1/df, log = FALSE)

pinvchisq(q, df, scale = 1/df, lower.tail = TRUE, log.p = FALSE)

qinvchisq(p, df, scale = 1/df, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{x}{vector of quantiles.}

\item{log}{logical; if \code{TRUE}, probabilities/densities \eqn{p}
    are returned as \eqn{log(p)}.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if \code{TRUE}, probabilities/densities \eqn{p}
    are returned as \eqn{log(p)}.}

\item{p}{vector of probabilities.}
}
\value{
\code{dinvchisq} gives the density, \code{pinvchisq} gives the distribution function,
\code{qinvchisq} gives the quantile function, and \code{rinvchisq}
generates random deviates.

Invalid arguments will result in return value \code{NaN}, with a warning.

The length of the result is determined by \code{n} for \code{rinvchisq},
and is the maximum of the lengths of the numerical arguments for the
other functions.

The numerical arguments other than \code{n} are recycled to the length
of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function, and random generation for
the (scaled) inverse chi-square distribution with \code{df} degrees of freedom and
optional parameter \code{scale}. The parameterization is consistent with Gelman et al. (2013). By default,
the \code{scale} is set to return results for the inverse chi-square distribution. If
\code{scale} is changed, then the results are returned for the scaled inverse chi-squared distribution.
}
\details{
If \code{scale} is omitted, it assumes the default value of \code{1/df}.

The scaled inverse chi-square distribution with parameters \code{df = n} and \code{scale = s^2} has density

\deqn{f(x)= (2^{-n/2}/Gamma(n/2)s^n x^(-n/2 - 1) e^(-n s^2/(2x))}

for \eqn{x \ge 0}, \eqn{n > 0} and \eqn{s > 0}. (Here \eqn{Gamma(n/2)} is the function implemented by R's
\code{gamma()} and defined in its help. Note that \eqn{n = 0} corresponds to the trivial distribution with all mass at point 0.)

The mean and variance are \eqn{E(X) = n/(n - 2) s^2} for \eqn{n > 2} and \eqn{Var(X) = 2n^2/(n-2)^2/(n-4) s^4} for \eqn{n>4}.

The cumulative hazard \eqn{H(t) = - log(1 - F(t))} is
\code{-pinvchisq(t, ..., lower = FALSE, log = TRUE)}
}
\examples{
x = 1:10
## InvChisquare(df = n) is a special case of InvGamma(shape = n/2, scale = 1/2)
n = 3
all.equal(dinvchisq(x, df = n), dinvgamma(x, shape = n/2, scale = 1/2))
all.equal(pinvchisq(x, df = n), pinvgamma(x, shape = n/2, scale = 1/2))

## InvChisquare(df = n, scale = s^2) is a special case of InvGamma(shape = n/2, scale = n/2 * s^2)
s = 1.7
all.equal(dinvchisq(x, df = n, scale = s^2), dinvgamma(x, shape = n/2, scale = n/2 * s^2))
all.equal(pinvchisq(x, df = n, scale = s^2), pinvgamma(x, shape = n/2, scale = n/2 * s^2))
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis, 3rd edition. CRC press.
}
\seealso{
\code{\link[stats]{Chisquare}}
}
