% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvGammaDist.R
\name{InvGammaDist}
\alias{InvGammaDist}
\alias{rinvgamma}
\alias{dinvgamma}
\alias{pinvgamma}
\alias{qinvgamma}
\title{The Inverse Gamma Distribution}
\usage{
rinvgamma(n, shape, scale = 1, rate = 1/scale)

dinvgamma(x, shape, scale = 1, rate = 1/scale, log = FALSE)

pinvgamma(
  q,
  shape,
  scale = 1,
  rate = 1/scale,
  lower.tail = TRUE,
  log.p = FALSE
)

qinvgamma(
  p,
  shape,
  scale = 1,
  rate = 1/scale,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{shape}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}

\item{x}{vector of quantiles.}

\item{log}{logical; if \code{TRUE}, probabilities/densities \eqn{p}
    are returned as \eqn{log(p)}.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if \code{TRUE}, probabilities/densities \eqn{p}
    are returned as \eqn{log(p)}.}

\item{p}{vector of probabilities.}
}
\value{
\code{dinvgamma} gives the density, \code{pinvgamma} gives the distribution function,
\code{qinvgamma} gives the quantile function, and \code{rinvgamma}
generates random deviates.

Invalid arguments will result in return value \code{NaN}, with a warning.

The length of the result is determined by \code{n} for \code{rinvgamma},
and is the maximum of the lengths of the numerical arguments for the
other functions.

The numerical arguments other than \code{n} are recycled to the length
of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function, and random generation for
the inverse gamma distribution with parameters \code{shape} and \code{scale}. Note that
the parameterization is consistent with Gelman et al. (2013) (see References) and
that \code{scale} precedes \code{rate} in the argument list, which is the
reverse of the \code{\link[stats]{GammaDist}} functions.
}
\details{
If \code{scale} is omitted, it assumes the default value of 1.

The inverse gamma distribution with parameters \code{shape = a} and \code{scale = s} has density

\deqn{f(x)= s^a/Gamma(a) x^(-a-1) e^-(s/x)}

for \eqn{x \ge 0}, \eqn{a > 0} and \eqn{s > 0}. (Here \eqn{Gamma(a)} is the function implemented by R's
\code{gamma()} and defined in its help. Note that \eqn{a = 0} corresponds to the trivial distribution with all mass at point 0.)

The mean and variance are \eqn{E(X) = s/(a - 1)} for \eqn{a>1} and \eqn{Var(X) = s^2/(a-1)^2/(a-2)} for \eqn{a>2}.

The cumulative hazard \eqn{H(t) = - log(1 - F(t))} is
\code{-pinvgamma(t, ..., lower = FALSE, log = TRUE)}
}
\examples{
-log(dinvgamma(1:4, shape = 1))
p <- (1:9)/10
pinvgamma(qinvgamma(p, shape = 2), shape = 2)
1 - 1/exp(qinvgamma(p, shape = 1))
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis, 3rd edition. CRC press.
}
\seealso{
\code{\link[stats]{GammaDist}}
}
