% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvWishart.R
\name{InvWishart}
\alias{InvWishart}
\alias{rinvwish}
\title{The Inverse Wishart Distribution}
\usage{
rinvwish(n, df, v)
}
\value{
The dimensionality of the result is determined by \code{n} for \code{rinvwish}.
The result will be an array of size \code{(nrow(v), nrow(v), n)}.
}
\description{
Random generation for the inverse Wishart distribution. The parameterization
is consistent with Gelman et al. (2013) (see References)
}
\examples{
Sigma = crossprod(matrix(rnorm(16), nrow = 4))
y = rinvwish(3, 4, Sigma)
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis, 3rd edition. CRC press.
}
\seealso{
\code{\link[stats]{rWishart}}
}
