% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{log.sum.exp}
\alias{log.sum.exp}
\title{Log-Sum-Exp Trick}
\usage{
\method{log}{sum.exp}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value: \eqn{\log(\sum \exp(x_i))}.
}
\description{
Computes the log of the sum of exponentials in a numerically stable way.
}
