% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logdiffexp}
\alias{logdiffexp}
\title{Log-difference-exp trick}
\usage{
logdiffexp(aa, bb)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Log-difference-exp trick
}
\details{
For a > b log(a-b) = log(exp(a) - exp(b))
}
