% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logsumexp}
\alias{logsumexp}
\title{Log-sum-exp trick}
\usage{
logsumexp(aa, bb)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Log-sum-exp trick
}
\details{
log(a+b) = log(exp(a) + exp(b))
}
