% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgesampling_utils.R
\name{lpf.poisson.gamma}
\alias{lpf.poisson.gamma}
\title{Un-normalized log posterior function corresponding to poisson-gamma.bug.R}
\usage{
lpf.poisson.gamma(sample.row, data)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Un-normalized log posterior function corresponding to poisson-gamma.bug.R
This is just a port of the bugs model code to pure R. For use a log_posterior
arguement in bridge_sampling
}
\details{
The function is XXXXXXX
}
