% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgesampling_utils.R
\name{lpf.poisson.gamma2}
\alias{lpf.poisson.gamma2}
\title{Vectorized un-normalized log posterior function corresponding to poisson-gamma.bug.R}
\usage{
lpf.poisson.gamma2(sample.row, data)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
More vectorized version of un-normalized log posterior function corresponding to
poisson-gamma.bug.R This is just a port of the bugs model code to pure R. For
use a log_posterior argument in bridge_sampling.
}
\details{
The function is XXXXXXX
}
