% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{mcmc.drawsmat}
\alias{mcmc.drawsmat}
\title{Extract Draws as Matrix}
\usage{
mcmc.drawsmat(draws, vars = NULL, ignore.lp = FALSE)
}
\arguments{
\item{draws}{A list of posterior samples.}

\item{vars}{Optional character vector of variables to extract.}

\item{ignore.lp}{Logical; if `TRUE`, removes `lp__`.}
}
\value{
A numeric matrix of posterior draws.
}
\description{
Returns posterior samples as a matrix for selected variables.
}
