% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.utils.R
\name{mcmc.summary}
\alias{mcmc.summary}
\title{Summarize Posterior Draws}
\usage{
mcmc.summary(
  fit_obj,
  draws = NULL,
  vars = NULL,
  ignore.lp = FALSE,
  digits = 3,
  PI.lvls = c(0.5, 0.9)
)
}
\arguments{
\item{fit_obj}{A fitted model object.}

\item{draws}{Optional. A list of posterior draws.}

\item{vars}{Optional character vector of variables to include.}

\item{ignore.lp}{Logical; if `TRUE`, removes `lp__` from summary.}

\item{digits}{Integer; number of decimal places to round.}

\item{PI.lvls}{Numeric vector of probability intervals.}
}
\value{
A data frame of summary statistics.
}
\description{
Provides summary statistics (mean, median, SD, intervals) of posterior samples.
}
