% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.file_utils.R
\name{open_mf}
\alias{open_mf}
\title{Open specified JAGS or Stan model file contained in bayesutils}
\usage{
open_mf(model.file.name)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Open the specified JAGS or Stan model file contained in bayesutils
}
