% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgesampling_utils.R
\name{parameter.bounds}
\alias{parameter.bounds}
\title{Set up paramerter bounds needed to use of JAGS output with bridgesampling library}
\usage{
parameter.bounds(param.info, params.alt.bounds.list = NULL)
}
\arguments{
\item{param.info}{An rjags object or a vector of parameter names}
}
\value{
The function will XX
}
\description{
Set up paramerter bounds needed to use of JAGS output with bridgesampling library.
If parameter bounds are -Inf to Inf, no need to specify that. Just specify bounds aternative
to this in a list with element names equal to the parameters needing alternative bounds.
}
\details{
The function is XXXXXXX
}
