% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{parameter.intervals}
\alias{parameter.intervals}
\title{HPDI and PI interval estimates from a parameter sample}
\usage{
parameter.intervals(
  param.sample,
  prob = 0.95,
  plotQ = F,
  main = "Histogram of param.samp",
  xlab = "param.samp",
  ylab,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
HPDI and PI interval estimates from a parameter sample
}
\details{
The function is a wrapper for coda and R interval functionality
}
