% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc_path.R
\name{plot_mcmc_path}
\alias{plot_mcmc_path}
\title{Plot path of MCMC chains}
\usage{
plot_mcmc_path(
  x,
  ncycles = 10,
  palette = "RdYlBu",
  xlim,
  ylim,
  xlab = "var 1",
  ylab = "var 2",
  type = "step",
  ...
)
}
\arguments{
\item{x}{A matrix with 2 columns or a list of 2 column matrices}

\item{ncycles}{The number of cycles to plot}

\item{palette}{a valid palette name (one of \code{palette.pals()}).
    The name is matched to the list of available palettes, ignoring
    upper vs. lower case, spaces, dashes, etc. in the matching.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{type}{Type of path to plot. \code{"step"} moves horizontally
and vertically for each variable step within the MCMC cycles. \code{"cycle"} plots
the variable movement for each completed MCMC cycle.}

\item{...}{Additional parameters for the \code{[graphics]{plot.default}} function.}
}
\value{
A plot.
}
\description{
This will produce a simple plot for visualizing the path of the variable steps (one variable
moving at a time) or cycles (the movement of the variables after completing
a cycle) for a list of MCMC chains stored as a list. The starting place of
the chain *the first row of each chain) is denoted by a point. See Details.
}
\details{
The function currently only works for chains with two variables stored as
a 2-column matrix or data frame. Each column contains the MCMC samples of
each variable.

The paths are distinguished by using different line colors. By default,
the \code{"RdYlBu"} palette from \code{\link[grDevices]{hcl.colors}} is used.
It is unlikely this color scheme will work well for more than 5 chains.
}
\examples{
x1 = cbind(rnorm(3), rchisq(3, df = 1))
x2 = cbind(rnorm(3), rchisq(3, df = 1))
x = list(x1, x2)
plot_mcmc_path(x)
# change point type
plot_mcmc_path(x, type = "cycle", pch = 20)
}
