% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqsim_plot.R
\name{qqsim_plot}
\alias{qqsim_plot}
\title{Plotting of empirical quantiles against quantiles of replicated data}
\usage{
qqsim_plot(
  x,
  sim_fun = function(x) rnorm(length(x), mean(x), sd(x)),
  Nsims = 100,
  Xhat = NULL,
  produce_plot = TRUE,
  return_data = FALSE
)
}
\arguments{
\item{x}{A vector of empirical data}

\item{sim_fun}{The function to simulate a vector of fake data. Defaults to sampling normal fake data with data mean and data std as sampling parameters.}

\item{Nsims}{The number of data simulation replicates.}

\item{Xhat}{A matrix of simulated data (Nsims times length(x)). When provided, the sim_fun argument is ignored.}

\item{produce_plot}{Should a plot actually be generated?}

\item{return_data}{If TRUE, underlying plotting data is returned.}
}
\value{
NULL or a list of plotting data.
}
\description{
Given some data vector x and a function to simulate fake data, qqsim_plot
produces a set of fake data replicates and plots empirical data quantiles
against the mean fake data quantiles. A 90 \% percentile interval is added to the plot to
account for sampling variation. Alternatively, a matrix of pre-generated fake
data can be provided.
}
