% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.file_utils.R
\name{read.model.file}
\alias{read.model.file}
\title{Read in specified JAGS or Stan model file contained in bayesutils as a string}
\usage{
read.model.file(model.file.name, model.path = NULL)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Read in specified JAGS or Stan model file contained in bayesutils as a string
}
