% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
rmvnorm(n = 1, mu, v, method = "eigen")
}
\arguments{
\item{n}{the number of samples required}

\item{mu}{a vector giving the mean of the random variables}

\item{v}{the covariance matrix of the random variables}

\item{method}{The method used to decompose \code{v}.
Valid options are \code{"chol"}, \code{"eigen"}, or
\code{"svd"}.}
}
\value{
An \eqn{m \times n} where \eqn{n} is the length of \code{mu}.
}
\description{
Simulate from a Multivariate Normal Distribution
}
\examples{
v = matrix(c(10,3,3,2), 2, 2)
y = rmvnorm(n = 4000, 1:2, v)
# should be close to 1:2
# rowMeans(y)
# should be close to v
# var(t(y))
}
