% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.utils.R
\name{round.format}
\alias{round.format}
\title{Formatter for genrating scientific Numeric Notations for Roundings with Zero-underflow}
\usage{
\method{round}{format}(x, digits = 2, threshold = 0.01)
}
\description{
Formatter for genrating scientific Numeric Notations for Roundings with Zero-underflow
}
