% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstanr.monitor.R
\name{sample.with.monitor}
\alias{sample.with.monitor}
\title{Start browser-based monitoring process for ongoing HMC sampling}
\usage{
sample.with.monitor(
  m,
  data,
  chains = 4,
  parallel_chains = 1,
  save_warmup = FALSE,
  refresh = NULL,
  init = NULL,
  iter_warmup = NULL,
  iter_sampling = NULL,
  sampling.monitor.dir = "sampling.monitor.dir",
  sampling.monitor.csv = paste0(sampling.monitor.dir, "/sampling.csv/"),
  basename.csv = "sampling",
  sampling.plots.dir = paste0(sampling.monitor.dir, "/sampling.plots"),
  ...
)
}
\arguments{
\item{m}{A CmdStanR model object.}

\item{data}{A list structure one would provide to the model and sample call through [m$sample(data = data)].}

\item{save_warmup}{See CmdStanR doc. If TRUE, the monitor will start plotting warumup draws as well (highlighted though to distringuish them from prod. draws). The monitor starts drawing after at least 100 production (or warumup) draws were generated.}

\item{refresh}{See CmdStanR doc.}

\item{init}{See CmdStanR doc.}

\item{iter_warmup}{See CmdStanR doc.}

\item{iter_sampling}{See CmdStanR doc.}

\item{sampling.monitor.dir}{The directory where all monitoring data is saved. Defaults to \emph{sampling.monitor.dir}}

\item{sampling.monitor.csv}{The subdirectory name below [sampling.monitor.dir] where draws CSV files are written to. Defaults to \emph{/sampling.csv/}.}

\item{basename.csv}{The draws CSV files prefix.}

\item{sampling.plots.dir}{The subdirectory name below [sampling.monitor.dir] where draws plot PNG files are written to. Defaults to \emph{/sampling.plots/}.}

\item{...}{Other CmdStanR-related options passed to [m$sample(...)].}

\item{pallalel_chains}{See CmdStanR doc. Mind that the monitor process takes away one effective CPU core.}
}
\value{
A CmdStanR fit object.
}
\description{
Invokes external R process to read HMC sampler data and provide simple browser-based trace plot visualizations, while the sampler is still doing it's thing. Helpful for
checking if divergent transitions or any other pathological chain behaviour are present without to wait for the sampler to finish. Requires the draws to be written locally to CSV. This functions effectively
wraps [m$sample], where [m] is a compiled CmdStan model, whereas an outside process is started to read generated CSV draws data, open your default webbrowser and
generate simple traceplot visualizations of obtained draws. The data and visualizations are refreshed continously. \bold{Warning: If the sampler is stopped, the monitor process keeps going and needs to be shut down manually as well.}
}
