% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.utils.R
\name{sdN}
\alias{sdN}
\title{Compute empirical Standard Deviation (without Bessel Adjustement)}
\usage{
sdN(x, na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\description{
Compute empirical Standard Deviation (without Bessel Adjustement)
}
