% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.utils.R
\name{se}
\alias{se}
\title{Compute Standard Error of the Arith. Mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Sample data}

\item{na.rm}{Ignore missing data}
}
\description{
Compute Standard Error of the Arith. Mean
}
