% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize a data vector}
\usage{
standardize(dvec, type = "regular")
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Standardize with mean centering and sd scaling or median centering and MAD scaling
}
\details{
XXXX
}
