% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update.bayesutils}
\alias{update.bayesutils}
\title{Standardize a data vector}
\usage{
\method{update}{bayesutils}()
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Standardize with mean centering and sd scaling or median centering and MAD scaling
}
\details{
Easily update the bayesutils library by installing the current version from the github site

Easily update the bayesutils library by installing the current version from the github site
}
