% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_bayesutils}
\alias{update_bayesutils}
\title{Easily update the bayesutils library by installing the current version from the github site}
\usage{
update_bayesutils()
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Easily update the bayesutils library by installing the current version from the github site
}
