% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.utils.R
\name{varN}
\alias{varN}
\title{Compute empirical Variance (without Bessel Adjustement)}
\usage{
varN(x, na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\description{
Compute empirical Variance (without Bessel Adjustement)
}
