% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.string_utils.R
\name{write.tmp.model.file}
\alias{write.tmp.model.file}
\title{XXXX}
\usage{
write.tmp.model.file(model.string)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
XXX
The function is XXXXXXX
}
