% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.file_utils.R
\name{write.tmp.model.file}
\alias{write.tmp.model.file}
\title{Write a JAGS or Stan model to file}
\usage{
write.tmp.model.file(model.string)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Write a JAGS or Stan model to a temp file. Model should be formatted as a string. Intended for
use with JAGS since rstan can already deal with model strings but JAGS needs a stand alone file.
}
